Relocate other debug related functions
